+++ packages/devextreme/js/__internal/ui/popup/popup_drag.ts
